Move existing code from local/writer pkg to new local/writer/volume pkg
