+++ internal/pkg/service/buffer/storage/level/local/writer/volume/file.go
@@ -1,4 +1,4 @@
-package writer
+package volume
 
 import (
 	"io"
