+++ internal/pkg/service/buffer/storage/level/local/writer/volume/volumes.go
@@ -1,4 +1,4 @@
-package writer
+package volume
 
 import (
 	"context"
