+++ internal/pkg/service/buffer/storage/level/local/writer/volume/writer_linux_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package writer
+package volume
 
 import (
 	"testing"
